Re-organize the profile scripts
During login, a DM will ignore the shell-specific profile scripts, and
load ~/.profile. Zsh ignores this file in favor of it's own .zprofile.

This leaves a disconnect which is best solved by ensuring environment is
configured in the generic ~/.profile script, and then that script gets
sourced for Zsh. This supports setting the environment for both types of
logins - whether using a DM or via interactive or login shell.